# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carshark",
        "version": "1.0"
    },
    "paths": {
...
        "/api/car_statuses": {
            "get": {
                "tags": [
                    "Статусы машин"
                ],
                "summary": "Получить список статусов машин",
                "description": "Получение списка всех статусов машин",
                "operationId": "/api/car_statuses(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает список статусов машин"
                    }
                }
            },
            "post": {
                "tags": [
                    "Статусы машин"
                ],
                "summary": "Добавить статус машины",
                "description": "Создание статуса машины",
                "operationId": "/api/car_statuses(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarStatusStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Создание нового статуса машины"
                    }
                }
            }
        },
        "/api/car_statuses/{id}": {
            "get": {
                "tags": [
                    "Статусы машин"
                ],
                "summary": "Получить статус машины",
                "description": "Получение статуса машины по идентификатору",
                "operationId": "/api/car_statuses{car_statuses}(GET)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор статуса машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает статус машины"
                    }
                }
            },
            "put": {
                "tags": [
                    "Статусы машин"
                ],
                "summary": "Редактировать статус машины",
                "description": "Редактирование статуса машины по идентификатору",
                "operationId": "/api/car_statuses(PUT)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор статуса машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarStatusStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Редактирование статуса машины"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Статусы машин"
                ],
                "summary": "Удалить событие машины",
                "description": "Удаление статуса машины по идентификатору",
                "operationId": "/api/car_statuses/{car_statuses}(DELETE)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор статуса машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удаление статуса машины"
                    }
                }
            }
        },
        "/api/operations": {
            "get": {
                "tags": [
                    "Операции"
                ],
                "summary": "Получить список операций",
                "description": "Получение списка всех операций",
                "operationId": "/api/operations(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает список операций"
                    }
                }
            },
            "post": {
                "tags": [
                    "Операции"
                ],
                "summary": "Добавить операцию",
                "description": "Создание операции",
                "operationId": "/api/operations(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OperationStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Создание новой операции"
                    }
                }
            }
        },
        "/api/operations/{id}": {
            "get": {
                "tags": [
                    "Операции"
                ],
                "summary": "Получить операцию",
                "description": "Получение операции по идентификатору",
                "operationId": "/api/operations/{operations}(GET)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор операции",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает операцию"
                    }
                }
            },
            "put": {
                "tags": [
                    "Операции"
                ],
                "summary": "Редактировать операцию",
                "description": "Редактирование операции по идентификатору",
                "operationId": "/api/operations(PUT)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор операции",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OperationStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Редактирование операции"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Операции"
                ],
                "summary": "Удалить операцию",
                "description": "Удаление операции по идентификатору",
                "operationId": "/api/operations/{operations}(DELETE)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор операции",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удаление операции"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CarStatusStoreRequest": {
                "required": [
                    "car_status"
                ],
                "properties": {
                    "car_status": {
                        "description": "Статус машины",
                        "type": "integer",
                        "example": "2"
                    }
                },
                "type": "object"
            },
...
            "OperationStoreRequest": {
                "required": [
                    "date",
                    "sum",
                    "renter_id"
                ],
                "properties": {
                    "date": {
                        "description": "Дата",
                        "type": "string",
                        "format": "date",
                        "example": "2019-12-19"
                    },
                    "sum": {
                        "description": "Сумма",
                        "type": "number",
                        "format": "float",
                        "example": "352.0"
                    },
                    "renter_id": {
                        "description": "id статуса арендатора",
                        "type": "integer",
                        "example": "2"
                    }
                },
                "type": "object"
            }
        }
    }
}